client/login: refactor dev fake role login buttons into seperate component
