+++ client/src/pages/Login/Login.tsx
@@ -1,32 +1,10 @@
-import { useAuth } from '../../contexts/auth/useAuth';
+import LoginDevButtons from './LoginDevButtons';
 import Logo from '/images/logo-main-black.png';
-import { useNavigate } from 'react-router-dom';
-import { roleLandingPages } from './roleLandingPages';
 
 export default function Login() {
-  const navigate = useNavigate();
-  const { user, setUser } = useAuth();
-
-  const handleLogin = (role: 'agent' | 'secretary' | 'doctor' | 'admin') => {
-    setUser({
-      id: 1,
-      firstname: 'John',
-      lastname: 'Doe',
-      email: 'test@test.com',
-      role: {
-        id: 1,
-        label: role
-      },
-      isArchived: false
-    });
-    navigate(roleLandingPages[role]);
-  };
-
   return (
     <section className="flex w-80 flex-col place-items-center place-self-center">
-      <h1 className="text-center font-medium">
-        Agenda Médical {user?.role.label}
-      </h1>
+      <h1 className="text-center font-medium">Agenda Médical</h1>
       <section className="flex w-full flex-col place-items-center gap-4">
         <img className="place-self-center" src={Logo} />
         <h2 className="text-center">Connexion</h2>
@@ -46,30 +24,7 @@ export default function Login() {
         <button disabled className="btn w-40 bg-primary">
           CONNEXION
         </button>
-        <button
-          onClick={() => handleLogin('agent')}
-          className="btn w-40 bg-primary"
-        >
-          Login Agent
-        </button>
-        <button
-          onClick={() => handleLogin('secretary')}
-          className="btn w-40 bg-primary"
-        >
-          Login Secretary
-        </button>
-        <button
-          onClick={() => handleLogin('doctor')}
-          className="btn w-40 bg-primary"
-        >
-          Login Doctor
-        </button>
-        <button
-          onClick={() => handleLogin('admin')}
-          className="btn w-40 bg-primary"
-        >
-          Login Admin
-        </button>
+        <LoginDevButtons />
       </section>
     </section>
   );
